Don't rely on unarchive download capabilities
So we had a problem with this part, I assume the corporate transparant proxying caused the download to stall very quickly (that's what we see when downloading ourselves, we think it's some anti-virus checking in the background).

And since unarchive does not have any timeout-related setting for downloading, and as the maintainer of unarchive, I don't think adding download-capabilities to unarchive was a smart thing to do. The fix is simple, download the archive using get_url instead, with a longer timeout. I chose 120 seconds, but less would probably do as well.